# backend/src/KapitelShelf.Api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "KapitelShelf.Api, Version=0.3.1.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Fetch all users.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user.",
        "requestBody": {
          "description": "The create user dto.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by the id.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user to get.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete a user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user to delete.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user to update.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated user.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{userId}/lastvisitedbooks": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get the last visited books of a user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/settings": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get the user settings of a user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSettingDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSettingDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSettingDTO"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a user setting of a user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The id of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The setting.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSettingDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSettingDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSettingDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the author id.",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "description": "Gets or sets or set the first name.",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Gets or sets the last name.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The Author DTO."
      },
      "BookDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the book id.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Gets or sets the title.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Gets or sets the description.",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "description": "Gets or sets the release date.",
            "format": "date-time",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "description": "Gets or sets the page number.",
            "format": "int32",
            "nullable": true
          },
          "series": {
            "$ref": "#/components/schemas/SeriesDTO"
          },
          "seriesNumber": {
            "type": "integer",
            "description": "Gets or sets the series number.",
            "format": "int32"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorDTO"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDTO"
            },
            "description": "Gets or sets the categories.",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDTO"
            },
            "description": "Gets or sets the tags.",
            "nullable": true
          },
          "cover": {
            "$ref": "#/components/schemas/FileInfoDTO"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDTO"
          }
        },
        "additionalProperties": false,
        "description": "The book dto."
      },
      "BookDTOPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDTO"
            },
            "description": "Gets or sets the items.",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "description": "Gets or sets the total number of items.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "The paginated result."
      },
      "CategoryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the category id.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or sets the name.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The category dto."
      },
...
      "CreateUserDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Gets or sets the username.",
            "nullable": true
          },
          "image": {
            "$ref": "#/components/schemas/ProfileImageTypeDTO"
          },
          "color": {
            "type": "string",
            "description": "Gets or sets the profile color.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The create dto for a user."
      },
      "FileInfoDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the fileinfo id.",
            "format": "uuid"
          },
          "filePath": {
            "type": "string",
            "description": "Gets or sets the file path.",
            "nullable": true
          },
          "fileSizeBytes": {
            "type": "integer",
            "description": "Gets or sets the file size in bytes.",
            "format": "int64"
          },
          "mimeType": {
            "type": "string",
            "description": "Gets or sets the mimetype.",
            "nullable": true
          },
          "sha256": {
            "type": "string",
            "description": "Gets or sets the sha256 checksum.",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "description": "Gets the filename.",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Gets or sets the fileinfo dto."
      },
...
      "LocationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the locaiton id.",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/LocationTypeDTO"
          },
          "url": {
            "type": "string",
            "description": "Gets or sets the url.",
            "nullable": true
          },
          "fileInfo": {
            "$ref": "#/components/schemas/FileInfoDTO"
          }
        },
        "additionalProperties": false,
        "description": "The location dto."
      },
      "LocationTypeDTO": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "description": "The location type enum.",
        "format": "int32"
      },
...
      "ProfileImageTypeDTO": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10
        ],
        "type": "integer",
        "description": "The profile image enum.",
        "format": "int32"
      },
      "SeriesDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the series id.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or sets the name.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Gets or sets the creation time.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Gets or sets the update time.",
            "format": "date-time"
          },
          "lastVolume": {
            "$ref": "#/components/schemas/BookDTO"
          },
          "totalBooks": {
            "type": "integer",
            "description": "Gets or sets the total number of books.",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The series dto."
      },
...
      "TagDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the tag id.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets or sets the name.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The tag dto."
      },
...
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the locaiton id.",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "description": "Gets or sets the user name.",
            "nullable": true
          },
          "image": {
            "$ref": "#/components/schemas/ProfileImageTypeDTO"
          },
          "color": {
            "type": "string",
            "description": "Gets or sets the profile color.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The user dto."
      },
      "UserSettingDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the user id.",
            "format": "uuid"
          },
          "key": {
            "type": "string",
            "description": "Gets or sets the setting key.",
            "nullable": true
          },
          "value": {
            "type": "string",
            "description": "Gets or sets the setting value.",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/UserSettingValueTypeDTO"
          }
        },
        "additionalProperties": false,
        "description": "The user dto."
      },
      "UserSettingValueTypeDTO": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "description": "The user setting value type enum.",
        "format": "int32"
      }
    }
  }
}